`commented.c` file finished
All of my comments start with

    //C:

I tried not to edit anything, just show the difference between the two
files.

I have no idea which of `test.c` or `code.c` everyone is using, but they
seem similar enough so I made a `diff.c` and `merge.c` to show the
difference and the git-style merge between the two, and `commented.c`
is a manual edit of the merge between `code.c` and `test.c`, with some
comments.

If this was not helpful, please let me know. If there's anything else I
can do, please let me know.